Move filterProtectedCustomvars() to the IdoCube class
