+++ library/Cube/Ido/IdoCube.php
@@ -3,6 +3,7 @@
 
 namespace Icinga\Module\Cube\Ido;
 
+use Icinga\Application\Config;
 use Icinga\Authentication\Auth;
 use Icinga\Data\Filter\Filter;
 use Icinga\Exception\ConfigurationError;
@@ -27,6 +28,13 @@ abstract class IdoCube extends DbCube
 
     /** @var MonitoringBackend */
     protected $backend;
+
+    /**
+     * Cache for {@link filterProtectedCustomvars()}
+     *
+     * @var string|null
+     */
+    protected $protectedCustomvars;
 
     /**
      * We can steal the DB connection directly from a Monitoring backend
@@ -129,5 +137,35 @@ abstract class IdoCube extends DbCube
         }
 
         return $restriction;
+    }
+
+    /**
+     * Return the given array without values matching the custom variables protected by the monitoring module
+     *
+     * @param   string[]    $customvars
+     *
+     * @return  string[]
+     */
+    protected function filterProtectedCustomvars(array $customvars)
+    {
+        if ($this->protectedCustomvars === null) {
+            $config = Config::module('monitoring')->get('security', 'protected_customvars');
+            $protectedCustomvars = array();
+
+            foreach (preg_split('~,~', $config, -1, PREG_SPLIT_NO_EMPTY) as $pattern) {
+                $regex = array();
+                foreach (explode('*', $pattern) as $literal) {
+                    $regex[] = preg_quote($literal, '/');
+                }
+
+                $protectedCustomvars[] = implode('.*', $regex);
+            }
+
+            $this->protectedCustomvars = empty($protectedCustomvars)
+                ? '/^$/'
+                : '/^(?:' . implode('|', $protectedCustomvars) . ')$/';
+        }
+
+        return preg_grep($this->protectedCustomvars, $customvars, PREG_GREP_INVERT);
     }
 }
