+++ library/Cube/Ido/IdoHostStatusCube.php
@@ -3,17 +3,8 @@
 
 namespace Icinga\Module\Cube\Ido;
 
-use Icinga\Application\Config;
-
 class IdoHostStatusCube extends IdoCube
 {
-    /**
-     * Cache for {@link filterProtectedCustomvars()}
-     *
-     * @var string|null
-     */
-    protected $protectedCustomvars;
-
     public function getRenderer()
     {
         return new IdoHostStatusCubeRenderer($this);
@@ -94,35 +85,5 @@ class IdoHostStatusCube extends IdoCube
         $select = $view->getQuery()->getSelectQuery();
 
         return $select;
-    }
-
-    /**
-     * Return the given array without values matching the custom variables protected by the monitoring module
-     *
-     * @param   string[]    $customvars
-     *
-     * @return  string[]
-     */
-    protected function filterProtectedCustomvars(array $customvars)
-    {
-        if ($this->protectedCustomvars === null) {
-            $config = Config::module('monitoring')->get('security', 'protected_customvars');
-            $protectedCustomvars = array();
-
-            foreach (preg_split('~,~', $config, -1, PREG_SPLIT_NO_EMPTY) as $pattern) {
-                $regex = array();
-                foreach (explode('*', $pattern) as $literal) {
-                    $regex[] = preg_quote($literal, '/');
-                }
-
-                $protectedCustomvars[] = implode('.*', $regex);
-            }
-
-            $this->protectedCustomvars = empty($protectedCustomvars)
-                ? '/^$/'
-                : '/^(?:' . implode('|', $protectedCustomvars) . ')$/';
-        }
-
-        return preg_grep($this->protectedCustomvars, $customvars, PREG_GREP_INVERT);
     }
 }
